// projects/magic-lib/src/lib/other/model/interface/magicGrid.interface.ts
export interface IValueGetterParams<T = any> {
    data: T;
    rowIndex: number;
    id: number;
    detailNode: IValueGetterParamsDetailNode;
    key?: string;
}
export interface IValueGetterParamsDetailNode {
    display: boolean;
}